requirements.txt
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
plotly==5.19.0
streamlit==1.32.0
matplotlib==3.8.3
tensorflow-cpu==2.15.0
pmdarima==2.0.4
nltk==3.8.1
statsmodels==0.14.1
prophet==1.1.5
